Don't rebind stdin/stdout during tests
This is required for this to run in an automated way, since the test run
now attempts to shove a sudo password into stdin. pytest and fabric have
disagreements over whether or not stdin is an open FD that can have
stuff put into it. For more information, google:
"UnsupportedOperation: redirected stdin is pseudofile, has no fileno()"