Print the list with newline delimiting
Print things to stderr when they would otherwise be printed to stdout.
This is a bit of a surgical action, because I specifically want to be
able to use the stdout of a mendel command in Jenkins to list the
versions, and don't want to have to ignore the 'Done.' text or the
various 'using config from...' output that would otherwise be output on
every initialization.

```
tweetlength % mendel list_nexus_versions --hide status 2>/dev/null
1.4
2.0.0
2.0.1
3.0.0
3.0.1
3.0.2
```